+++ intasend/src/lib.rs
@@ -9,7 +9,7 @@ pub use self::intasend::checkout::{
 pub use self::intasend::collection::{
     Collection, MpesaStkPushRequest, MpesaStkPushResponse, StkPushStatusRequest, StkPushStatusResponse,
 };
-pub use self::intasend::payouts::{Payout, PayoutRequest, Payouts};
+pub use self::intasend::payouts::{Payout, PayoutTransaction, PayoutRequest, Payouts};
 pub use self::intasend::refunds::{RefundRequest, Refunds, RefundResponse};
 pub use self::intasend::wallets::{FundCheckoutRequest, Wallet, Wallets};
 pub use self::intasend::{Currency, Intasend, Transaction};