+++ intasend/src/lib.rs
@@ -12,7 +12,7 @@ pub use self::intasend::collection::{
 pub use self::intasend::payouts::{Payout, PayoutTransaction, PayoutRequest, Payouts};
 pub use self::intasend::refunds::{RefundRequest, Refunds, RefundResponse};
 pub use self::intasend::wallets::{FundCheckoutRequest, Wallet, Wallets};
-pub use self::intasend::{Currency, Intasend, Transaction};
+pub use self::intasend::{Currency, Intasend, Transaction, PayoutProvider};
 
 #[cfg(test)]
 mod tests {